Struct bytes::alloc::Pool

pub struct Pool {
    // some fields omitted
}

Methods

impl Pool

fn with_capacity(cap: usize, buf_len: usize) -> Pool

Constructs a new Pool with with specified capacity such that each buffer in the pool has a length of buf_len.

fn capacity(&self) -> usize

Returns the number of buffers that the Pool holds.

fn buffer_len(&self) -> usize

Returns the size of buffers allocated by the pool

fn new_byte_buf(&self) -> Option<MutByteBuf>

Returns a new ByteBuf backed by a buffer from the pool. If the pool is depleted, None is returned.

fn new_append_buf(&self) -> Option<AppendBuf>