Module bytes::buf::ext

Extra utilities for Buf and BufMut types.

Structs

Chain
A Chain sequences two buffers.

Limit
A BufMut adapter which limits the amount of bytes that can be written to an underlying buffer.

Reader
A Buf adapter which implements io::Read for the inner value.

Take
A Buf adapter which limits the bytes read from an underlying buffer.

Writer
A BufMut adapter which implements io::Write for the inner value.

Traits

BufExt
Extra methods for implementations of Buf.

BufMutExt
Extra methods for implementations of BufMut.