Struct bytes::buf::Chain

pub struct Chain<T, U> { /* fields omitted */ }

A Chain sequences two buffers.

Chain is an adapter that links two underlying buffers and provides a continous view across both buffers. It is able to sequence either immutable buffers (Buf values) or mutable buffers (BufMut values).

This struct is generally created by calling Buf::chain. Please see that function's documentation for more detail.

Examples

use bytes::{Bytes, Buf, IntoBuf};
use bytes::buf::Chain;

let buf = Bytes::from(&b"hello "[..]).into_buf()
            .chain(Bytes::from(&b"world"[..]));

let full: Bytes = buf.collect();
assert_eq!(full[..], b"hello world"[..]);

Methods

impl<T, U> Chain<T, U>

pub fn new(a: T, b: U) -> Chain<T, U>

Creates a new Chain sequencing the provided values.

Examples

use bytes::BytesMut;
use bytes::buf::Chain;

let buf = Chain::new(
    BytesMut::with_capacity(1024),
    BytesMut::with_capacity(1024));

// Use the chained buffer

pub fn first_ref(&self) -> &T

Gets a reference to the first underlying Buf.

Examples

use bytes::{Bytes, Buf, IntoBuf};

let buf = Bytes::from(&b"hello"[..]).into_buf()
            .chain(Bytes::from(&b"world"[..]));

assert_eq!(buf.first_ref()[..], b"hello"[..]);

pub fn first_mut(&mut self) -> &mut T

Gets a mutable reference to the first underlying Buf.

Examples

use bytes::{Bytes, Buf, IntoBuf};

let mut buf = Bytes::from(&b"hello "[..]).into_buf()
                .chain(Bytes::from(&b"world"[..]));

buf.first_mut().advance(1);

let full: Bytes = buf.collect();
assert_eq!(full[..], b"ello world"[..]);

pub fn last_ref(&self) -> &U

Gets a reference to the last underlying Buf.

Examples

use bytes::{Bytes, Buf, IntoBuf};

let buf = Bytes::from(&b"hello"[..]).into_buf()
            .chain(Bytes::from(&b"world"[..]));

assert_eq!(buf.last_ref()[..], b"world"[..]);

pub fn last_mut(&mut self) -> &mut U

Gets a mutable reference to the last underlying Buf.

Examples

use bytes::{Bytes, Buf, IntoBuf};

let mut buf = Bytes::from(&b"hello "[..]).into_buf()
                .chain(Bytes::from(&b"world"[..]));

buf.last_mut().advance(1);

let full: Bytes = buf.collect();
assert_eq!(full[..], b"hello orld"[..]);

pub fn into_inner(self) -> (T, U)

Consumes this Chain, returning the underlying values.

Examples

use bytes::{Bytes, Buf};

let chain = Bytes::from(&b"hello"[..])
            .chain(Bytes::from(&b"world"[..]));

let (first, last) = chain.into_inner();
assert_eq!(first[..], b"hello"[..]);
assert_eq!(last[..], b"world"[..]);