bytes::buf

Trait IntoBuf

pub trait IntoBuf {
    type Buf: Buf;

    // Required method
    fn into_buf(self) -> Self::Buf;
}

Conversion into a Buf

An IntoBuf implementation defines how to convert a value into a Buf. This is common for types that represent byte storage of some kind. IntoBuf may be implemented directly for types or on references for those types.

Examples

use bytes::{Buf, IntoBuf, BigEndian};

let bytes = b"\x00\x01hello world";
let mut buf = bytes.into_buf();

assert_eq!(1, buf.get_u16::<BigEndian>());

let mut rest = [0; 11];
buf.copy_to_slice(&mut rest);

assert_eq!(b"hello world", &rest);

Required Associated Types

type Buf: Buf

The Buf type that self is being converted into

Required Methods

fn into_buf(self) -> Self::Buf

Creates a Buf from a value.

Examples

use bytes::{Buf, IntoBuf, BigEndian};

let bytes = b"\x00\x01hello world";
let mut buf = bytes.into_buf();

assert_eq!(1, buf.get_u16::<BigEndian>());

let mut rest = [0; 11];
buf.copy_to_slice(&mut rest);

assert_eq!(b"hello world", &rest);