Trait bytes::buf::MutBufExt

pub trait MutBufExt {
    fn write<S: Source>(&mut self, src: S) -> Result<usize, S::Error>;
}

An extension trait providing extra functions applicable to all MutBuf values.

Required methods

fn write<S: Source>(&mut self, src: S) -> Result<usize, S::Error>

Write bytes from the given source into the current MutBuf and advance the cursor by the number of bytes written.

Loading content...