Struct bytes::bytes::Inner

#[repr(C)]struct Inner {
    arc: AtomicPtr<Shared>,
    ptr: *mut u8,
    len: usize,
    cap: usize,
}

Fields
arc: AtomicPtr<Shared>ptr: *mut u8len: usizecap: usize
Implementations

impl Inner

fn from_static(bytes: &'static [u8]) -> Inner

fn from_vec(src: Vec<u8>) -> Inner

fn with_capacity(capacity: usize) -> Inner

fn as_ref(&self) -> &[u8]

Return a slice for the handle's view into the shared buffer

fn as_mut(&mut self) -> &mut [u8]

Return a mutable slice for the handle's view into the shared buffer

unsafe fn as_raw(&mut self) -> &mut [u8]

Return a mutable slice for the handle's view into the shared buffer including potentially uninitialized bytes.

fn put_u8(&mut self, n: u8)

Insert a byte into the next slot and advance the len by 1.

fn len(&self) -> usize

unsafe fn inline_ptr(&self) -> *mut u8

Pointer to the start of the inline buffer

fn inline_len(&self) -> usize

fn set_inline_len(&mut self, len: usize)

Set the length of the inline buffer. This is done by writing to the least significant byte of the arc field.

unsafe fn set_len(&mut self, len: usize)

slice.

fn is_empty(&self) -> bool

fn capacity(&self) -> usize

fn split_off(&mut self, at: usize) -> Inner

fn split_to(&mut self, at: usize) -> Inner

fn truncate(&mut self, len: usize)

fn resize(&mut self, new_len: usize, value: u8)

unsafe fn set_start(&mut self, start: usize)

unsafe fn set_end(&mut self, end: usize)

fn is_mut_safe(&mut self) -> bool

Checks if it is safe to mutate the memory

unsafe fn shallow_clone(&self, mut_self: bool) -> Inner

Increments the ref count. This should only be done if it is known that it can be done safely. As such, this fn is not public, instead other fns will use this one while maintaining the guarantees. Parameter mut_self should only be set to true if caller holds &mut self reference.

"Safely" is defined as not exposing two BytesMut values that point to the same byte window.

This function is thread safe.

unsafe fn shallow_clone_sync(&self, mut_self: bool) -> Inner

unsafe fn shallow_clone_arc(&self, arc: *mut Shared) -> Inner

unsafe fn shallow_clone_vec(&self, arc: usize, mut_self: bool) -> Inner

fn reserve(&mut self, additional: usize)

fn is_inline(&self) -> bool

Returns true if the buffer is stored inline

fn is_inline_or_static(&self) -> bool

fn is_shared(&mut self) -> bool

Used for debug_assert statements. &mut is used to guarantee that it is safe to check VEC_KIND

fn is_static(&mut self) -> bool

Used for debug_assert statements

fn kind(&self) -> usize

fn uncoordinated_get_vec_pos(&mut self) -> (usize, usize)

fn uncoordinated_set_vec_pos(&mut self, pos: usize, prev: usize)