bytes

Module bytes_mut

Structs

BytesMut

A unique reference to a contiguous slice of memory.

Shared 🔒

Constants

KIND_ARC 🔒

KIND_MASK 🔒

KIND_VEC 🔒

MAX_ORIGINAL_CAPACITY_WIDTH 🔒

MAX_VEC_POS 🔒

MIN_ORIGINAL_CAPACITY_WIDTH 🔒

NOT_VEC_POS_MASK 🔒

ORIGINAL_CAPACITY_MASK 🔒

ORIGINAL_CAPACITY_OFFSET 🔒

PTR_WIDTH 🔒

VEC_POS_OFFSET 🔒

Statics

SHARED_VTABLE 🔒

Functions

_split_must_use 🔒

_split_off_must_use 🔒

_split_to_must_use 🔒

increment_shared 🔒 ⚠

invalid_ptr 🔒

Returns a dangling pointer with the given address. This is used to store integer data in pointer fields.

original_capacity_from_repr 🔒

original_capacity_to_repr 🔒

rebuild_vec 🔒 ⚠

release_shared 🔒 ⚠

shared_v_clone 🔒 ⚠

shared_v_drop 🔒 ⚠

shared_v_is_unique 🔒 ⚠

shared_v_to_mut 🔒 ⚠

shared_v_to_vec 🔒 ⚠

vptr 🔒