Module bytes::str

Structs

Bytes
A specialized ByteStr box.

Rope
An immutable sequence of bytes formed by concatenation of other ByteStr values, without copying the data in the pieces. The concatenation is represented as a tree whose leaf nodes are each a Bytes value.

RopeBuf

SeqByteStr

SmallByteStr

SmallByteStrBuf

Traits

ByteStr
An immutable sequence of bytes. Operations will not mutate the original value. Since only immutable access is permitted, operations do not require copying (though, sometimes copying will happen as an optimization).

ToBytes