Struct bytes::str::Bytes

pub struct Bytes { /* fields omitted */ }

A specialized ByteStr box.

Methods

impl Bytes

pub fn from_slice(bytes: &[u8]) -> Bytes

pub fn of<B: ByteStr>(bytes: B) -> Bytes

pub fn empty() -> Bytes

pub fn downcast_ref<'a, B: ByteStr>(&'a self) -> Option<&'a B>

If the underlying ByteStr is of type B, returns a reference to it otherwise None.

pub fn try_unwrap<B: ByteStr>(self) -> Result<B, Bytes>

If the underlying ByteStr is of type B, returns the unwraped value, otherwise, returns the original Bytes as Err.