Struct bytes::AppendBuf

pub struct AppendBuf {
    // some fields omitted
}

A Buf backed by a contiguous region of memory.

This buffer can only be written to once. Byte strings (immutable views) can be created at any time, not just when the writing is complete.

Methods

impl AppendBuf

fn with_capacity(capacity: u32) -> AppendBuf

unsafe fn from_mem_ref(mem: MemRef, cap: u32, pos: u32) -> AppendBuf

fn len(&self) -> usize

fn capacity(&self) -> usize

fn bytes(&self) -> &[u8]

fn shift(&self, n: usize) -> Bytes

fn drop(&self, n: usize)

fn slice(&self, begin: usize, end: usize) -> Bytes