Struct bytes::BlockBuf

pub struct BlockBuf {
    // some fields omitted
}

Append only buffer backed by a chain of AppendBuf buffers.

Each AppendBuf block is of a fixed size and allocated on demand. This makes the total capacity of a BlockBuf potentially much larger than what is currently allocated.

Methods

impl BlockBuf

fn new(max_blocks: usize, block_size: usize) -> BlockBuf

Create BlockBuf

fn len(&self) -> usize

Returns the number of buffered bytes

fn is_empty(&self) -> bool

Returns true if there are no buffered bytes

fn buf(&self) -> BlockBufCursor

Returns a Buf for the currently buffered bytes.

fn shift(&mut self, n: usize) -> Bytes

Consumes n buffered bytes, returning them as an immutable Bytes value.

Panics

Panics if n is greater than the number of buffered bytes.

fn drop(&mut self, n: usize)

Drop the first n buffered bytes

Panics

Panics if n is greater than the number of buffered bytes.

fn is_compact(&mut self) -> bool

fn compact(&mut self)

Moves all buffered bytes into a single block.

Panics

Panics if the buffered bytes cannot fit in a single block.

fn bytes(&self) -> Option<&[u8]>

Return byte slice if bytes are in sequential memory