Struct bytes::Bytes

pub struct Bytes {
    // some fields omitted
}

Methods

impl Bytes

fn empty() -> Bytes

Return an empty Bytes

unsafe fn from_mem_ref(mem: MemRef, pos: u32, len: u32) -> Bytes

Creates a new Bytes from a MemRef, an offset, and a length.

This function is unsafe as there are no guarantees that the given arguments are valid.

fn buf(&self) -> BytesBuf

fn is_empty(&self) -> bool

fn len(&self) -> usize

fn concat(&self, other: &Bytes) -> Bytes

fn slice(&self, begin: usize, end: usize) -> Bytes

Returns a new ByteStr value containing the byte range between begin (inclusive) and end (exclusive)

fn slice_from(&self, begin: usize) -> Bytes

Returns a new ByteStr value containing the byte range starting from begin (inclusive) to the end of the byte str.

Equivalent to bytes.slice(begin, bytes.len())

fn slice_to(&self, end: usize) -> Bytes

Returns a new ByteStr value containing the byte range from the start up to end (exclusive).

Equivalent to bytes.slice(0, end)