Struct bytes::SliceBuf

pub struct SliceBuf<T> { /* fields omitted */ }

A Buf backed by a contiguous region of memory.

This Buf is better suited for cases where there is a clear delineation between reading and writing.

Methods

impl<T: AsRef<[u8]>> SliceBuf<T>

fn new(mem: T) -> SliceBuf<T>

Creates a new SliceBuf wrapping the provided slice

fn capacity(&self) -> usize

Return the number of bytes the buffer can contain

fn position(&self) -> usize

Return the read cursor position

fn set_position(&mut self, position: usize)

Set the read cursor position

fn len(&self) -> usize

Return the number of buffered bytes

fn is_empty(&self) -> bool

Returns true if the buffer contains no unread bytes

fn clear(&mut self)

Clears the buffer, removing any written data