Struct bytes::TakeMut

pub struct TakeMut<T> { /* fields omitted */ }

A buffer adapter which limits the bytes written from an underlying value.

Methods

impl<T> TakeMut<T>

fn into_inner(self) -> T

Consumes this TakeMut, returning the underlying value.

fn get_ref(&self) -> &T

Gets a reference to the underlying value in this TakeMut.

fn get_mut(&mut self) -> &mut T

Gets a mutable reference to the underlying value in this TakeMut.

fn limit(&self) -> usize

Returns the maximum number of bytes that are made available from the underlying value.

fn set_limit(&mut self, lim: usize)

Sets the maximum number of bytes that are made available from the underlying value.