Trait bytes::IntoBuf

pub trait IntoBuf {
    type Buf: Buf;
    fn into_buf(self) -> Self::Buf;
}

Conversion into a Buf

Usually, IntoBuf is implemented on references of types and not directly on the types themselves. For example, IntoBuf is implemented for &'a Vec<u8> and not Vec<u8> directly.

Associated Types

type Buf: Buf

The Buf type that self is being converted into

Required Methods

fn into_buf(self) -> Self::Buf

Creates a Buf from a value.