Crate bytes

Reexports

pub use buf::{ByteBuf, MutByteBuf, RingBuf, ROByteBuf, SliceBuf, MutSliceBuf, Take};

pub use str::{Bytes, Rope, RopeBuf, SeqByteStr, SmallByteStr, SmallByteStrBuf};

Modules

alloc

traits
All traits are re-exported here to allow glob imports.

Enums

BufError

Traits

Buf
A trait for values that provide sequential read access to bytes.

BufExt
An extension trait providing extra functions applicable to all Buf values.

ByteStr
An immutable sequence of bytes. Operations will not mutate the original value. Since only immutable access is permitted, operations do not require copying (though, sometimes copying will happen as an optimization).

MutBuf
A trait for values that provide sequential write access to bytes.

MutBufExt
An extension trait providing extra functions applicable to all MutBuf values.

Sink
A value that reads bytes from a Buf into itself

Source
A value that writes bytes from itself into a MutBuf.

ToBytes