Crate bytes

Provides abstractions for working with bytes.

Structs

ByteBuf
A buffer backed by BytesMut

Bytes
A reference counted slice of bytes.

BytesMut
A unique reference to a slice of bytes.

Reader
Adapts a Buf to the io::Read trait

SliceBuf
A Buf backed by a contiguous region of memory.

Take
A buffer adapter which limits the bytes read from an underlying value.

TakeMut
A buffer adapter which limits the bytes written from an underlying value.

Writer
Adapts a BufMut to the io::Write trait

Traits

Buf
A trait for values that provide sequential read access to bytes.

BufMut
A trait for values that provide sequential write access to bytes.

IntoBuf
Conversion into a Buf

Sink
A value that copies bytes from a Buf into itself

Source
A value that writes bytes from itself into a BufMut.