Crate bytes

Re-exports

pub use buf::Buf;

pub use buf::BufExt;

pub use buf::MutBuf;

pub use buf::MutBufExt;

pub use buf::Source;

pub use buf::Sink;

pub use str::ByteStr;

pub use str::ToBytes;

Modules

alloc

buf

str

traits
All traits are re-exported here to allow glob imports.

Structs

ByteBuf
A Buf backed by a contiguous region of memory.

Bytes
A specialized ByteStr box.

MutByteBuf

MutSliceBuf

ROByteBuf
Same as ByteBuf but cannot be flipped to a MutByteBuf.

RingBuf
Buf backed by a continous chunk of memory. Maintains a read cursor and a write cursor. When reads and writes reach the end of the allocated buffer, wraps around to the start.

Rope
An immutable sequence of bytes formed by concatenation of other ByteStr values, without copying the data in the pieces. The concatenation is represented as a tree whose leaf nodes are each a Bytes value.

RopeBuf

SeqByteStr

SliceBuf

SmallByteStr

SmallByteStrBuf

Take

Enums

BufError