Crate bytes

Modules

traits

Structs

ByteBuf
A Buf backed by a contiguous region of memory.

Bytes

MutByteBuf

MutSliceBuf

ROByteBuf
Same as ByteBuf but cannot be flipped to a MutByteBuf.

RingBuf
Buf backed by a continous chunk of memory. Maintains a read cursor and a write cursor. When reads and writes reach the end of the allocated buffer, wraps around to the start.

RingBufReader

RingBufWriter

Rope
An immutable sequence of bytes formed by concatenation of other ByteStr values, without copying the data in the pieces. The concatenation is represented as a tree whose leaf nodes are each a Bytes value.

SeqByteStr

SliceBuf

SmallByteStr

SmallByteStrBuf

Enums

BufError

Traits

Buf
A trait for values that provide random and sequential access to bytes.

BufExt

ByteStr

MutBuf

MutBufExt

Sink
An value that reads bytes from a Buf into itself

Source

Type Definitions

BufResult