Module bytes::buf

Utilities for working with buffers.

A buffer is any structure that contains a sequence of bytes. The bytes may or may not be stored in contiguous memory. This module contains traits used to abstract over buffers as well as utilities for working with buffer types.

Buf, BufMut

These are the two foundational traits for abstractly working with buffers. They can be thought as iterators for byte structures. They offer additional performance over Iterator by providing an API optimized for byte slices.

See Buf and BufMut for more details.

Re-exports

pub use self::ext::BufExt;

pub use self::ext::BufMutExt;

Modules

ext
Extra utilities for Buf and BufMut types.

Structs

IntoIter
Iterator over the bytes contained by the buffer.

IoSliceMut
A buffer type used for readv.

Traits

Buf
Read bytes from a buffer.

BufMut
A trait for values that provide sequential write access to bytes.