Module bytes::buf

Structs

ByteBuf
A Buf backed by a contiguous region of memory.

MutByteBuf

MutSliceBuf

ROByteBuf
Same as ByteBuf but cannot be flipped to a MutByteBuf.

RingBuf
Buf backed by a continous chunk of memory. Maintains a read cursor and a write cursor. When reads and writes reach the end of the allocated buffer, wraps around to the start.

SliceBuf

Take

Traits

Buf
A trait for values that provide sequential read access to bytes.

BufExt
An extension trait providing extra functions applicable to all Buf values.

MutBuf
A trait for values that provide sequential write access to bytes.

MutBufExt
An extension trait providing extra functions applicable to all MutBuf values.

Sink
A value that reads bytes from a Buf into itself

Source
A value that writes bytes from itself into a MutBuf.