Struct bytes::buf::Chain

pub struct Chain<T, U> { /* fields omitted */ }

A Chain sequences two buffers.

Chain is an adapter that links two underlying buffers and provides a continuous view across both buffers. It is able to sequence either immutable buffers (Buf values) or mutable buffers (BufMut values).

This struct is generally created by calling Buf::chain. Please see that function’s documentation for more detail.

Examples

use bytes::{Bytes, Buf};

let mut buf = (&b"hello "[..])
    .chain(&b"world"[..]);

let full: Bytes = buf.copy_to_bytes(11);
assert_eq!(full[..], b"hello world"[..]);

Implementations

impl<T, U> Chain<T, U>

pub fn first_ref(&self) -> &T

Gets a reference to the first underlying Buf.

Examples

use bytes::Buf;

let buf = (&b"hello"[..])
    .chain(&b"world"[..]);

assert_eq!(buf.first_ref()[..], b"hello"[..]);

pub fn first_mut(&mut self) -> &mut T

Gets a mutable reference to the first underlying Buf.

Examples

use bytes::Buf;

let mut buf = (&b"hello"[..])
    .chain(&b"world"[..]);

buf.first_mut().advance(1);

let full = buf.copy_to_bytes(9);
assert_eq!(full, b"elloworld"[..]);

pub fn last_ref(&self) -> &U

Gets a reference to the last underlying Buf.

Examples

use bytes::Buf;

let buf = (&b"hello"[..])
    .chain(&b"world"[..]);

assert_eq!(buf.last_ref()[..], b"world"[..]);

pub fn last_mut(&mut self) -> &mut U

Gets a mutable reference to the last underlying Buf.

Examples

use bytes::Buf;

let mut buf = (&b"hello "[..])
    .chain(&b"world"[..]);

buf.last_mut().advance(1);

let full = buf.copy_to_bytes(10);
assert_eq!(full, b"hello orld"[..]);

pub fn into_inner(self) -> (T, U)

Consumes this Chain, returning the underlying values.

Examples

use bytes::Buf;

let chain = (&b"hello"[..])
    .chain(&b"world"[..]);

let (first, last) = chain.into_inner();
assert_eq!(first[..], b"hello"[..]);
assert_eq!(last[..], b"world"[..]);