Struct bytes::buf::Reader

pub struct Reader<B> { /* fields omitted */ }

A Buf adapter which implements io::Read for the inner value.

This struct is generally created by calling reader() on Buf. See documentation of reader() for more details.

Methods

impl<B: Buf> Reader<B>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_ref(&self) -> &B

Gets a reference to the underlying Buf.

It is inadvisable to directly read from the underlying Buf.

Examples

use bytes::Buf;
use std::io::{self, Cursor};

let mut buf = Cursor::new(b"hello world").reader();

assert_eq!(0, buf.get_ref().position());

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get_mut(&mut self) -> &mut B

Gets a mutable reference to the underlying Buf.

It is inadvisable to directly read from the underlying Buf.

Examples

use bytes::Buf;
use std::io::{self, Cursor};

let mut buf = Cursor::new(b"hello world").reader();
let mut dst = vec![];

buf.get_mut().set_position(2);
io::copy(&mut buf, &mut dst).unwrap();

assert_eq!(*dst, b"llo world"[..]);

pub fn into_inner(self) -> B

Consumes this Reader, returning the underlying value.

Examples

use bytes::Buf;
use std::io::{self, Cursor};

let mut buf = Cursor::new(b"hello world").reader();
let mut dst = vec![];

io::copy(&mut buf, &mut dst).unwrap();

let buf = buf.into_inner();
assert_eq!(0, buf.remaining());