Module bytes::bytes_mut

Structs

BytesMut

A unique reference to a contiguous slice of memory.

Shared ğŸ”’

Constants

KIND_ARC ğŸ”’

KIND_MASK ğŸ”’

KIND_VEC ğŸ”’

MAX_ORIGINAL_CAPACITY_WIDTH ğŸ”’

MAX_VEC_POS ğŸ”’

MIN_ORIGINAL_CAPACITY_WIDTH ğŸ”’

NOT_VEC_POS_MASK ğŸ”’

ORIGINAL_CAPACITY_MASK ğŸ”’

ORIGINAL_CAPACITY_OFFSET ğŸ”’

PTR_WIDTH ğŸ”’

VEC_POS_OFFSET ğŸ”’

Statics

SHARED_VTABLE ğŸ”’

Functions

_split_must_use ğŸ”’

_split_off_must_use ğŸ”’

_split_to_must_use ğŸ”’

increment_shared ğŸ”’ âš

invalid_ptr ğŸ”’

Returns a dangling pointer with the given address. This is used to store integer data in pointer fields.

original_capacity_from_repr ğŸ”’

original_capacity_to_repr ğŸ”’

rebuild_vec ğŸ”’ âš

release_shared ğŸ”’ âš

shared_v_clone ğŸ”’ âš

shared_v_drop ğŸ”’ âš

shared_v_is_unique ğŸ”’ âš

shared_v_to_mut ğŸ”’ âš

shared_v_to_vec ğŸ”’ âš

vptr ğŸ”’