Struct bytes::ByteBuf

pub struct ByteBuf { /* fields omitted */ }

A Buf backed by a contiguous region of memory.

This Buf is better suited for cases where there is a clear delineation between reading and writing.

Methods

impl ByteBuf

Important traits for ByteBuf
impl Read for ByteBuf

pub fn from_slice(bytes: &[u8]) -> ByteBuf

Create a new ByteBuf by copying the contents of the given slice.

Important traits for MutByteBuf
impl Write for MutByteBuf

pub fn mut_with_capacity(capacity: usize) -> MutByteBuf

Important traits for ByteBuf
impl Read for ByteBuf

pub fn none() -> ByteBuf

Important traits for ByteBuf
impl Read for ByteBuf

pub unsafe fn from_mem_ref(mem: MemRef, cap: u32, pos: u32, lim: u32) -> ByteBuf

pub fn capacity(&self) -> usize

Important traits for MutByteBuf
impl Write for MutByteBuf

pub fn flip(self) -> MutByteBuf

Important traits for MutByteBuf
impl Write for MutByteBuf

pub fn resume(self) -> MutByteBuf

Flips the buffer back to mutable, resetting the write position to the byte after the previous write.

pub fn read_slice(&mut self, dst: &mut [u8]) -> usize

pub fn to_seq_byte_str(self) -> SeqByteStr

pub fn to_bytes(self) -> Bytes

pub fn mark(&mut self)

Marks the current read location.

Together with reset, this can be used to read from a section of the buffer multiple times. The marked location will be cleared when the buffer is flipped.

pub fn reset(&mut self)

Resets the read position to the previously marked position.

Together with mark, this can be used to read from a section of the buffer multiple times.

Panics

This method will panic if no mark has been set.